requirements.txt
pytesseract
google-generativeai
streamlit
opencv-python
mediapipe
numpy
Pillow

